revert the changes of moving common data models to interface file to avoid breaking change
