+++ packages/teams-js/src/private/externalAppCardActions.ts
@@ -4,7 +4,7 @@ import { ApiName, ApiVersionNumber, getApiVersionTag } from '../internal/telemet
 import { AppId } from '../public';
 import { errorNotSupportedOnPlatform, FrameContexts } from '../public/constants';
 import { runtime } from '../public/runtime';
-import { ActionOpenUrlError, ActionSubmitError, IAdaptiveCardActionSubmit } from './interfaces';
+import { ExternalAppErrorCode } from './constants';
 
 /**
  * v2 APIs telemetry file: All of APIs in this capability file should send out API version v2 ONLY
@@ -29,6 +29,62 @@ export namespace externalAppCardActions {
     DeepLinkOther = 'DeepLinkOther',
     DeepLinkStageView = 'DeepLinkStageView',
     GenericUrl = 'GenericUrl',
+  }
+
+  /**
+   * @beta
+   * @hidden
+   * Error that can be thrown from IExternalAppCardActionService.handleActionOpenUrl
+   * and IExternalAppCardActionForCEAService.handleActionOpenUrl
+   *
+   * @internal
+   * Limited to Microsoft-internal use
+   */
+  export interface ActionOpenUrlError {
+    errorCode: ActionOpenUrlErrorCode;
+    message?: string;
+  }
+
+  /**
+   * @beta
+   * @hidden
+   * Error codes that can be thrown from IExternalAppCardActionService.handleActionOpenUrl
+   * and IExternalAppCardActionForCEAService.handleActionOpenUrl
+   *
+   * @internal
+   * Limited to Microsoft-internal use
+   */
+  export enum ActionOpenUrlErrorCode {
+    INTERNAL_ERROR = 'INTERNAL_ERROR', // Generic error
+    INVALID_LINK = 'INVALID_LINK', // Deep link is invalid
+    NOT_SUPPORTED = 'NOT_SUPPORTED', // Deep link is not supported
+  }
+
+  /**
+   * @beta
+   * @hidden
+   * The payload that is used when executing an Adaptive Card Action.Submit
+   *
+   * @internal
+   * Limited to Microsoft-internal use
+   */
+  export interface IAdaptiveCardActionSubmit {
+    id: string;
+    data: string | Record<string, unknown>;
+  }
+
+  /**
+   * @beta
+   * @hidden
+   * Error that can be thrown from IExternalAppCardActionService.handleActionSubmit
+   * and IExternalAppCardActionForCEAService.handleActionSubmit
+   *
+   * @internal
+   * Limited to Microsoft-internal use
+   */
+  export interface ActionSubmitError {
+    errorCode: ExternalAppErrorCode;
+    message?: string;
   }
 
   /**
