+++ packages/teams-js/test/private/externalAppCardActionsForCEA.spec.ts
@@ -2,8 +2,8 @@ import { errorLibraryNotInitialized } from '../../src/internal/constants';
 import { GlobalVars } from '../../src/internal/globalVars';
 import { ApiName } from '../../src/internal/telemetry';
 import { ExternalAppErrorCode } from '../../src/private/constants';
+import { externalAppCardActions } from '../../src/private/externalAppCardActions';
 import { externalAppCardActionsForCEA } from '../../src/private/externalAppCardActionsForCEA';
-import { ActionOpenUrlErrorCode, ActionOpenUrlType } from '../../src/private/interfaces';
 import { AppId, FrameContexts } from '../../src/public';
 import { app } from '../../src/public/app';
 import { errorNotSupportedOnPlatform } from '../../src/public/constants';
@@ -123,10 +123,10 @@ describe('externalAppCardActionsForCEA', () => {
     const allowedFrameContexts = [FrameContexts.content];
     const testUrl = new URL('https://example.com');
     const testError = {
-      errorCode: ActionOpenUrlErrorCode.INTERNAL_ERROR,
+      errorCode: externalAppCardActions.ActionOpenUrlErrorCode.INTERNAL_ERROR,
       message: 'testMessage',
     };
-    const testResponse = ActionOpenUrlType.DeepLinkDialog;
+    const testResponse = externalAppCardActions.ActionOpenUrlType.DeepLinkDialog;
 
     it('should not allow calls before initialization', async () => {
       expect.assertions(1);
